+++ src/app/admin/components/index.ts
@@ -1,0 +1,4 @@
+export * from './product-item/product-item.component';
+export * from './product-form/product-form.component';
+export * from './products/products.component';
+export * from './orders/orders.component';