+++ src/app/app.module.ts
@@ -3,26 +3,30 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+import { AdminModule } from './admin/admin.module';
 import { CartModule } from './cart/cart.module';
 import { CoreModule } from './core/core.module';
 import { FirstComponent } from './first/first.component';
 import { OrdersModule } from './orders/orders.module';
 import { ProductsModule } from './products/products.module';
 import { SharedModule } from './shared/shared.module';
+import { LoginComponent } from './login/login.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponent
+    FirstComponent,
+    LoginComponent
   ],
   imports: [
+    AdminModule,
     BrowserModule,
-    AppRoutingModule,
     CartModule,
     CoreModule,
     OrdersModule,
     ProductsModule,
-    SharedModule
+    SharedModule,
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]