+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
+
 import { CartProductsService } from 'src/app/cart/services/cart-products.service';
+import { UserRoleService } from 'src/app/core/services/user-role.service';
 import { ProductModel } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
 
@@ -13,7 +16,12 @@ export class ProductListComponent implements OnInit {
 
   products$: Observable<ProductModel[]>;
 
-  constructor(private productsService: ProductsService, private cartProductsService: CartProductsService) { }
+  constructor(
+    public userRoleService: UserRoleService,
+    private router: Router, 
+    private productsService: ProductsService, 
+    private cartProductsService: CartProductsService
+  ) { }
 
   ngOnInit(): void {
     this.getProducts();
@@ -28,5 +36,10 @@ export class ProductListComponent implements OnInit {
       const product = val.find(x => x.id === id);
       this.cartProductsService.addProduct(product);
     })
+  }
+
+  onShowProduct(product: ProductModel): void {
+    const link = ['/product', product.id];
+    this.router.navigate(link);
   }
 }