+++ src/app/products/products.module.ts
@@ -1,16 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { ProductComponent } from './components/product/product.component';
-import { ProductListComponent } from './components/product-list/product-list.component';
+import { ProductComponent } from './components';
+import { ProductsRoutingModule } from './products-routing.module';
 
 @NgModule({
-  declarations: [ProductComponent, ProductListComponent],
+  declarations: [ProductComponent, ProductsRoutingModule.components],
   imports: [
-    CommonModule
-  ],
-  exports: [
-    ProductListComponent
+    CommonModule,
+    ProductsRoutingModule
   ]
 })
 export class ProductsModule { }